pyproject.toml
[project]
name = "tadas-butkus-reprice-relize"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "filelock>=3.25.2",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "numpy>=2.4.3",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "schedule>=1.2.2",
]
